Add express routing boilerplate
