+++ src/api/auth.ts
@@ -1,0 +1,10 @@
+import express from 'express';
+
+const router = express.Router();
+
+router.post('/login', (req, res) => {
+  // TODO
+  res.status(204).end();
+});
+
+export default router;
